adding datapacks back to 5.1.0 until UI has filter logic
Datapacks need to remain because the datapack category is
hardcoded in the UI in older versions of CDAP. But the UI also
does not have logic to hide the category if it exists but there
are no packages for that specific CDAP version.

Until that logic is there, datapacks must be visible.